[JDF-316] - Refactoring packages.
