+++ picketlink-authentication-idm-multi-tennancy/src/main/java/org/jboss/as/quickstarts/picketlink/idm/totp/jsf/LoginController.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.as.quickstarts.picketlink.authentication.totp.jsf;
+package org.jboss.as.quickstarts.picketlink.idm.totp.jsf;
 
 import javax.enterprise.context.RequestScoped;
 import javax.faces.application.FacesMessage;
